feat: iniciado configuracoes de estruturas do projeto com base no sketch da UI
