+++ src/components/header.tsx
@@ -1,0 +1,5 @@
+const Header = () => {
+  return <header className="bg-black text-white text-lg h-12">Header</header>;
+};
+
+export default Header;
