+++ src/components/page-layout.tsx
@@ -1,0 +1,12 @@
+import Header from "./header";
+
+const PageLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <div className="w-screen h-screen flex flex-col">
+      <Header />
+      {children}
+    </div>
+  );
+};
+
+export default PageLayout;
