+++ src/components/parameters-selector.tsx
@@ -1,0 +1,5 @@
+const ParametersSelector = () => {
+  return <div className="bg-blue-400 h-full w-48">Parameters Selector</div>;
+};
+
+export default ParametersSelector;
